# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce de productos orgánicos",
        "description": "Documentación para la API de ecommerce de productos orgánicos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend-raices.vercel.app/api/api",
            "description": "Servidor de producción"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener todos los roles",
                "description": "Devuelve una lista de roles. Solo los administradores (rol id=1) pueden acceder a esta operación.",
                "operationId": "6e28bdec08d7292bde33180ca5928718",
                "responses": {
                    "200": {
                        "description": "Lista de roles"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "description": "Crea un nuevo rol en el sistema. Solo los administradores (rol id=1) pueden realizar esta operación.",
                "operationId": "b47f38c8385567ad7a473d82fb6ebf63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Role"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener un rol por ID",
                "description": "Devuelve los detalles de un rol específico. Solo los administradores (rol id=1) pueden realizar esta operación.",
                "operationId": "a4b5c5aa1ac224a279b2060bab05733c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del rol"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "description": "Elimina un rol específico del sistema. Solo los administradores (rol id=1) pueden realizar esta operación.",
                "operationId": "f5117dc2943b621bdb53820fcc72b457",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Rol eliminado exitosamente"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar un rol existente",
                "description": "Actualiza los detalles de un rol específico. Solo los administradores (rol id=1) pueden realizar esta operación.",
                "operationId": "852042c9cece9741fddbf2a40bc2535e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Role"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado exitosamente"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    },
                    "400": {
                        "description": "Solicitud inválida"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Role": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Usuario administrador"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}